requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.2
fsspec==2025.10.0
h11==0.16.0
httptools==0.7.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lap==0.5.12
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
mysql-connector-python==9.4.0
networkx==3.2.1
numpy==2.0.2
omegaconf==2.3.0
opencv-python==4.10.0.84
optional==0.0.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qtgui==0.0.1
reportlab==4.4.7
requests==2.32.5
router==0.1
scipy==1.13.1
six==1.17.0
starlette==0.49.3
sympy==1.13.1
tomli==2.4.0
tools==1.0.17
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.250
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.39.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.5
zipp==3.23.0
